pyproject.toml
[tool.poetry]
name = "pyzettle"
version = "0.1.0"
description = "Python package to access the Zettle APIs"
authors = ["Alex Baugh <132655501+alex-baugh@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
